+++ src/lib/content.js
@@ -1,23 +1,13 @@
 const autoAssumeLastRole = new AutoAssumeLastRole();
 
 function extendIAMFormList() {
-  var csrf, list = elById('awsc-username-menu-recent-roles');
-  if (list) {
-    var firstForm = list.querySelector('#awsc-recent-role-0 form');
-    csrf = firstForm['csrf'].value;
-  } else {
-    list = generateEmptyRoleList();
-    csrf = '';
-  }
-
   const lastRoleKey = autoAssumeLastRole.createKey();
 
   chrome.storage.sync.get([
     'profiles', 'profiles_1', 'profiles_2', 'profiles_3', 'profiles_4',
-    'hidesHistory', 'hidesAccountId', 'showOnlyMatchingRoles',
+    'hidesAccountId', 'showOnlyMatchingRoles',
     'autoAssumeLastRole', lastRoleKey
   ], function(data) {
-    var hidesHistory = data.hidesHistory || false;
     var hidesAccountId = data.hidesAccountId || false;
     var showOnlyMatchingRoles = data.showOnlyMatchingRoles || false;
     autoAssumeLastRole.enabled = data.autoAssumeLastRole || false;
@@ -26,7 +16,7 @@ function extendIAMFormList() {
       const dps = new DataProfilesSplitter();
       const profiles = dps.profilesFromDataSet(data);
 
-      loadProfiles(new ProfileSet(profiles, showOnlyMatchingRoles), list, csrf, hidesHistory, hidesAccountId);
+      loadProfiles(new ProfileSet(profiles, showOnlyMatchingRoles), list);
       attachColorLine(profiles);
     }
     // console.log("Last role from '"+vlastRoleKey+"' was '"+lastRole+"'");
@@ -34,45 +24,7 @@ function extendIAMFormList() {
   });
 }
 
-function generateEmptyRoleList() {
-  var divLbl = document.createElement('div');
-  divLbl.id = 'awsc-recent-roles-label';
-  divLbl.textContent = 'Role List:';
-  var ul = document.createElement('ul');
-  ul.id = 'awsc-username-menu-recent-roles';
-
-  var parentEl = elById('awsc-login-account-section');
-  parentEl.appendChild(divLbl);
-  parentEl.appendChild(ul);
-
-  var script = document.createElement('script');
-  script.src = chrome.extension.getURL('/js/csrf-setter.js');
-  parentEl.appendChild(script);
-  return ul;
-}
-
-function replaceRedirectURI(form, profile) {
-  if (!profile.region) return false;
-
-  const destRegion = profile.region;
-  var redirectUri = decodeURIComponent(form.redirect_uri.value);
-  const md = redirectUri.match(/region=([a-z\-1-9]+)/);
-  if (md) {
-    const currentRegion = md[1];
-    if (currentRegion !== destRegion) {
-      redirectUri = redirectUri.replace(new RegExp(currentRegion, 'g'), destRegion);
-      if (currentRegion === 'us-east-1') {
-        redirectUri = redirectUri.replace('://', `://${destRegion}.`);
-      } else if (destRegion === 'us-east-1') {
-        redirectUri = redirectUri.replace(/:\/\/[^.]+\./, '://');
-      }
-    }
-    form.redirect_uri.value = encodeURIComponent(redirectUri);
-  }  
-}
-
 function hookBeforeSwitch(form, profile) {
-  replaceRedirectURI(form, profile);
   autoAssumeLastRole.save(profile);
   return true;
 }
@@ -82,88 +34,8 @@ function hookBeforeExit() {
   return true;
 }
 
-function loadProfiles(profileSet, list, csrf, hidesHistory, hidesAccountId) {
+function loadProfiles(profileSet, list) {
   var recentNames = [];
-
-  if (hidesHistory) {
-    var fc = list.firstChild;
-    while (fc) {
-      list.removeChild(fc);
-      fc = list.firstChild;
-    }
-
-    var label = elById('awsc-recent-roles-label');
-    if (label) {
-      label.textContent = label.textContent.replace('History', 'List');
-    }
-  } else {
-    var li = list.firstElementChild;
-    while (li) {
-      input = li.querySelector('input[type="submit"]');
-      let name = input.value;
-      name = name.replace(/\s+\|\s+\d+$/, '');
-      if (profileSet.excludedNames.includes(name)) {
-        const nextLi = li.nextElementSibling;
-        list.removeChild(li);
-        li = nextLi;
-      } else {
-        const form = li.querySelector('form');
-        form.dataset.aesrProfile = name;
-        input.style = 'white-space:pre';
-        recentNames.push(name);
-        li = li.nextElementSibling;
-      }
-    }
-  }
-
-  const redirectUri = encodeURIComponent(window.location.href);
-  profileSet.destProfiles.forEach(function(item) {
-    var name = item.profile;
-    if (recentNames.indexOf(name) !== -1) return true;
-    if (!hidesAccountId) name += '  |  ' + item.aws_account_id;
-
-    var color = item.color || 'aaaaaa';
-    let actionHost;
-    if (window.location.host.endsWith('.amazonaws-us-gov.com')) {
-        actionHost = 'signin.amazonaws-us-gov.com';
-    } else if (window.location.host.endsWith('.amazonaws.cn')) {
-        actionHost = 'signin.amazonaws.cn';
-    } else {
-        actionHost = 'signin.aws.amazon.com';
-    }
-    if (!item.image) {
-        list.insertAdjacentHTML('beforeend', Sanitizer.escapeHTML`<li>
-         <form action="https://${actionHost}/switchrole" method="POST" target="_top" data-aesr-profile="${item.profile}">
-          <input type="hidden" name="action" value="switchFromBasis">
-          <input type="hidden" name="src" value="nav">
-          <input type="hidden" name="roleName" value="${item.role_name}">
-          <input type="hidden" name="account" value="${item.aws_account_id}">
-          <input type="hidden" name="mfaNeeded" value="0">
-          <input type="hidden" name="color" value="${color}">
-          <input type="hidden" name="csrf" value="${csrf}">
-          <input type="hidden" name="redirect_uri" value="${redirectUri}">
-          <label for="awsc-recent-role-switch-0" class="awsc-role-color" style="background-color: #${color};">&nbsp;</label>
-          <input type="submit" class="awsc-role-submit awsc-role-display-name" name="displayName" value="${name}"
-                title="${item.role_name}@${item.aws_account_id}" style="white-space:pre"></form>
-        </li>`);
-    } else {
-        list.insertAdjacentHTML('beforeend', Sanitizer.escapeHTML`<li>
-         <form action="https://${actionHost}/switchrole" method="POST" target="_top" data-aesr-profile="${item.profile}">
-          <input type="hidden" name="action" value="switchFromBasis">
-          <input type="hidden" name="src" value="nav">
-          <input type="hidden" name="roleName" value="${item.role_name}">
-          <input type="hidden" name="account" value="${item.aws_account_id}">
-          <input type="hidden" name="mfaNeeded" value="0">
-          <input type="hidden" name="color" value="${color}">
-          <input type="hidden" name="csrf" value="${csrf}">
-          <input type="hidden" name="redirect_uri" value="${redirectUri}">
-          <label for="awsc-recent-role-switch-0" class="awsc-role-color"><img src=${item.image.replace(/"/g, '')} style="margin-top: -1px; margin-left: -1px; width: 17px; height: 17px"></label>
-          <input type="submit" class="awsc-role-submit awsc-role-display-name" name="displayName" value="${name}"
-                title="${item.role_name}@${item.aws_account_id}" style="white-space:pre"></form>
-        </li>`);
-
-    }
-  });
 
   Array.from(list.querySelectorAll('form')).forEach(form => {
     form.onsubmit = function(e) {
@@ -179,42 +51,6 @@ function loadProfiles(profileSet, list, csrf, hidesHistory, hidesAccountId) {
       hookBeforeExit(this);
     });
   }
-
-  // Place role filter textinput
-  var AWSR_firstForm = null;
-
-  document.getElementById('awsc-recent-roles-label').insertAdjacentHTML('beforeend', '<input id="AESR_RoleFilter" type="text" placeholder="Filter by profile name" style="border:1px solid #ccc;border-radius:3px;font-size:13px;margin-left:0.25em;max-width:20ex;padding:0.4ex">');
-
-  document.getElementById('AESR_RoleFilter').onkeyup = function(e) {
-    const str = this.value;
-    if (e.keyCode === 13) {
-      if (AWSR_firstForm) {
-        AWSR_firstForm.querySelector('input[type="submit"]').click()
-      }
-    } else {
-      const lis = Array.from(document.querySelectorAll('#awsc-username-menu-recent-roles > li'));
-      let firstHitLi = null;
-      lis.forEach(li => {
-        const profileName = li.firstElementChild.querySelector("input[name='displayName']").value.toLowerCase();
-        const hit = str ? profileName.indexOf(str) > -1 : false;
-        const shown = str ? hit : true;
-        li.style.display = shown ? 'block' : 'none';
-        li.style.background = null;
-        if (hit && firstHitLi === null) firstHitLi = li;
-      });
-
-      if (firstHitLi) {
-        firstHitLi.style.background = '#f0f9ff';
-        AWSR_firstForm = firstHitLi.querySelector('form');
-      } else {
-        AWSR_firstForm = null;
-      }
-    }
-  }
-
-  document.getElementById('nav-usernameMenu').addEventListener('click', () => {
-    document.getElementById('AESR_RoleFilter').focus()
-  })
 }
 
 function attachColorLine(profiles) {
@@ -257,4 +93,3 @@ function needsInvertForeColorByBack(color) {
 
   return (0.299 * r + 0.587 * g + 0.114 * b) / 255 < 0.5;
 }
-